Separate command utility and class tests
