+++ core/src/main/java/io/spine/core/Commands.java
@@ -20,10 +20,8 @@
 
 package io.spine.core;
 
-import com.google.protobuf.Duration;
 import com.google.protobuf.Message;
 import com.google.protobuf.util.Timestamps;
-import io.spine.annotation.Internal;
 import io.spine.base.CommandMessage;
 import io.spine.protobuf.Messages;
 import io.spine.string.Stringifier;
@@ -31,10 +29,7 @@ import io.spine.string.StringifierRegistry;
 
 import java.util.List;
 
-import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.core.CommandContext.Schedule;
-import static io.spine.validate.Validate.isNotDefault;
 
 /**
  * Client-side utilities for working with commands.
@@ -53,20 +48,6 @@ public final class Commands {
     }
 
     /**
-     * Extracts the message from the passed {@code Command} instance.
-     *
-     * @param command a command to extract a message from
-     * @return an unpacked message
-     * @deprecated use {@link Command#enclosedMessage()}
-     */
-    @Deprecated
-    public static CommandMessage getMessage(Command command) {
-        checkNotNull(command);
-        CommandMessage result = command.enclosedMessage();
-        return result;
-    }
-
-    /**
      * Extracts a command message if the passed instance is a {@link Command} object or
      * {@link com.google.protobuf.Any Any}, otherwise returns the passed message.
      */
@@ -80,34 +61,6 @@ public final class Commands {
     }
 
     /**
-     * Obtains a tenant ID from the command.
-     *
-     * @deprecated Please use {@link Command#tenant()}.
-     */
-    @Deprecated
-    public static TenantId tenantOf(Command command) {
-        checkNotNull(command);
-        TenantId result = command.tenant();
-        return result;
-    }
-
-    /**
-     * Obtains a {@link TenantId} from the {@link CommandContext}.
-     *
-     * <p>The {@link CommandContext} is accessible from the {@link Event} if the {@code Event} was 
-     * created as a result of some command or its rejection. This makes the {@code CommandContext}
-     * a valid {@code TenantId} source inside of the {@code Event}.
-     *
-     * @deprecated Please use {@link Command#tenant()}
-     */
-    @Deprecated
-    @Internal
-    public static TenantId tenantOf(CommandContext context) {
-        return context.getActorContext()
-                      .getTenantId();
-    }
-
-    /**
      * Sorts the command given command request list by command timestamp value.
      *
      * @param commands the command list to sort
@@ -115,26 +68,6 @@ public final class Commands {
     public static void sort(List<Command> commands) {
         checkNotNull(commands);
         commands.sort((c1, c2) -> Timestamps.compare(c1.time(), c2.time()));
-    }
-
-    /**
-     * Checks if the command is scheduled to be delivered later.
-     *
-     * @param command a command to check
-     * @return {@code true} if the command context has a scheduling option set,
-     * {@code false} otherwise
-     */
-    public static boolean isScheduled(Command command) {
-        checkNotNull(command);
-        Schedule schedule = command.context()
-                                   .getSchedule();
-        Duration delay = schedule.getDelay();
-        if (isNotDefault(delay)) {
-            checkArgument(delay.getSeconds() > 0,
-                          "Command delay seconds must be a positive value.");
-            return true;
-        }
-        return false;
     }
 
     /**
