+++ server/src/main/java/io/spine/server/commandbus/CommandScheduler.java
@@ -38,7 +38,6 @@ import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.Sets.newHashSet;
 import static com.google.protobuf.util.Timestamps.checkValid;
 import static io.spine.base.Time.currentTime;
-import static io.spine.core.Commands.isScheduled;
 import static io.spine.server.bus.Buses.acknowledge;
 import static java.util.Optional.empty;
 
@@ -80,7 +79,7 @@ public abstract class CommandScheduler implements BusFilter<CommandEnvelope> {
     @Override
     public Optional<Ack> accept(CommandEnvelope envelope) {
         Command command = envelope.command();
-        if (isScheduled(command)) {
+        if (command.isScheduled()) {
             schedule(envelope.command());
             return Optional.of(acknowledge(envelope.id()));
         }
